Give each canvas a reference to the canvas manager it uses.
Right now, there is an implicit, but not explict, link in the code between the canvas manager and the canvases, so any algorithm that tries to trace widget dependencies misses that the canvas manager is associated with each canvas. This makes that association explicit (as a bonus, allows a person to create multiple canvas managers if they so desire, I suppose).